refactor(content): introduce WKC adapter layout and relocate trivial ports
Establishes the Well-Known Components (WKC) layout for the content server.
Components are split into adapters/ (I/O) and logic/ (rules/orchestration);
each component lives in its own folder with component.ts, types.ts,
index.ts (and optional errors.ts). Controllers are not components.

Adds project CLAUDE.md describing the convention and pointing to the
dcl-wkc-components skill.

Relocates ports that were already shaped like adapters into
src/adapters/<name>/{component,types,index}.ts:

- ports/clock -> adapters/clock
- ports/sequecuentialTaskExecutor -> adapters/sequential-task-executor (typo)
- ports/denylist -> adapters/denylist
- ports/fileWriter -> adapters/content-file-writer
- ports/deployRateLimiterComponent -> adapters/deploy-rate-limiter
- ports/dao-servers-getter -> adapters/dao-client
- ports/system-properties + logic/database-queries/system-properties-queries
  -> adapters/system-properties (single adapter owning SQL and the typed-key
  facade)

Drops ports/contentFileWriter (no consumers).

No behavior changes; all importers updated, tsc clean, unit suite green
(213 tests).
