+++ content/src/adapters/clock/component.ts
@@ -1,0 +1,7 @@
+import { Clock } from './types'
+
+export function createClock(): Clock {
+  return {
+    now: Date.now
+  }
+}
