+++ content/src/adapters/clock/index.ts
@@ -1,0 +1,2 @@
+export { createClock } from './component'
+export type { Clock } from './types'
