+++ content/src/adapters/clock/types.ts
@@ -1,0 +1,3 @@
+export interface Clock {
+  now(): number
+}
