+++ content/src/adapters/content-file-writer/index.ts
@@ -1,0 +1,2 @@
+export { createFileWriter } from './component'
+export type { IFile } from './types'
