+++ content/src/adapters/content-file-writer/types.ts
@@ -1,0 +1,7 @@
+export type IFile = {
+  filePath: string
+  appendDebounced: (buffer: string) => Promise<void>
+  close: () => Promise<void>
+  delete: () => Promise<void>
+  store: () => Promise<string>
+}
