+++ content/src/adapters/dao-client/index.ts
@@ -1,0 +1,2 @@
+export { createDAOComponent, createCustomDAOComponent } from './component'
+export type { DAOComponent } from './types'
