+++ content/src/adapters/dao-client/types.ts
@@ -1,0 +1,6 @@
+import { CatalystServerInfo } from '@dcl/catalyst-contracts'
+
+export interface DAOComponent {
+  getAllContentServers(): Promise<CatalystServerInfo[]>
+  getAllServers(): Promise<CatalystServerInfo[]>
+}
