+++ content/src/adapters/denylist/index.ts
@@ -1,0 +1,2 @@
+export { createDenylist } from './component'
+export type { Denylist } from './types'
