+++ content/src/adapters/denylist/types.ts
@@ -1,0 +1,5 @@
+export type Denylist = {
+  isDenylisted: (id: string) => boolean
+  start?: () => Promise<void>
+  stop?: () => Promise<void>
+}
