+++ content/src/adapters/deploy-rate-limiter/index.ts
@@ -1,0 +1,2 @@
+export { createDeployRateLimiter } from './component'
+export type { IDeployRateLimiterComponent, DeploymentRateLimitConfig } from './types'
