+++ content/src/adapters/sequential-task-executor/index.ts
@@ -1,0 +1,2 @@
+export { createSequentialTaskExecutor, sequentialJobMetrics } from './component'
+export type { ISequentialTaskExecutorComponent } from './types'
