+++ content/src/adapters/sequential-task-executor/types.ts
@@ -1,0 +1,6 @@
+export type ISequentialTaskExecutorComponent = {
+  /**
+   * Runs sequential jobs with a max concurrency of 1 per jobName.
+   */
+  run<T>(jobName: string, fn: () => Promise<T>): Promise<T>
+}
