+++ content/src/adapters/system-properties/index.ts
@@ -1,0 +1,2 @@
+export { createSystemProperties, SYSTEM_PROPERTIES } from './component'
+export type { SystemProperties, SystemProperty } from './types'
