+++ content/src/components.ts
@@ -19,20 +19,20 @@ import { splitByCommaTrimAndRemoveEmptyElements } from './logic/config-helpers'
 import { metricsDeclaration } from './metrics'
 import { createMigrationExecutor } from './migrations/migration-executor'
 import { createActiveEntitiesComponent } from './ports/activeEntities'
-import { createClock } from './ports/clock'
-import { createCustomDAOComponent, createDAOComponent } from './ports/dao-servers-getter'
-import { createDenylist } from './ports/denylist'
-import { createDeployRateLimiter } from './ports/deployRateLimiterComponent'
+import { createClock } from './adapters/clock'
+import { createCustomDAOComponent, createDAOComponent } from './adapters/dao-client'
+import { createDenylist } from './adapters/denylist'
+import { createDeployRateLimiter } from './adapters/deploy-rate-limiter'
 import { createDeployedEntitiesBloomFilter } from './ports/deployedEntitiesBloomFilter'
 import { createDeployer } from './ports/deployer'
 import { createFailedDeployments } from './ports/failedDeployments'
 import { createDatabaseComponent } from './ports/postgres'
 import { createProcessedSnapshotStorage } from './ports/processedSnapshotStorage'
-import { createSequentialTaskExecutor } from './ports/sequecuentialTaskExecutor'
+import { createSequentialTaskExecutor } from './adapters/sequential-task-executor'
 import { createSnapshotGenerator } from './ports/snapshotGenerator'
 import { createSnapshotStorage } from './ports/snapshotStorage'
 import { createSynchronizationState } from './ports/synchronizationState'
-import { createSystemProperties } from './ports/system-properties'
+import { createSystemProperties } from './adapters/system-properties'
 import { ContentAuthenticator } from './service/auth/Authenticator'
 import { GarbageCollectionManager } from './service/garbage-collection/GarbageCollectionManager'
 import { PointerManager } from './service/pointers/PointerManager'
