--- content/src/ports/clock.ts
@@ -1,9 +1,0 @@
-export interface Clock {
-  now(): number
-}
-
-export function createClock() {
-  return {
-    now: Date.now
-  }
-}
