+++ content/test/mocks/deploy-rate-limiter-mock.ts
@@ -1,4 +1,4 @@
-import { IDeployRateLimiterComponent } from '../../src/ports/deployRateLimiterComponent'
+import { IDeployRateLimiterComponent } from '../../src/adapters/deploy-rate-limiter'
 
 export function createNoOpDeployRateLimiter(): IDeployRateLimiterComponent {
   return {
