+++ content/test/mocks/sequential-task-executor-componen-mock.ts
@@ -1,4 +1,4 @@
-import { ISequentialTaskExecutorComponent } from '../../src/ports/sequecuentialTaskExecutor'
+import { ISequentialTaskExecutorComponent } from '../../src/adapters/sequential-task-executor'
 
 export const createMockedSequentialTaskExecutorComponent = (
   overrides: Partial<jest.Mocked<ISequentialTaskExecutorComponent>> = {}
