+++ content/test/unit/ports/activeEntities.spec.ts
@@ -11,8 +11,8 @@ import { DEFAULT_ENTITIES_CACHE_SIZE, Environment, EnvironmentConfig } from '../
 import * as deployments from '../../../src/logic/deployments'
 import { metricsDeclaration } from '../../../src/metrics'
 import { ActiveEntities, createActiveEntitiesComponent } from '../../../src/ports/activeEntities'
-import { createClock } from '../../../src/ports/clock'
-import { Denylist } from '../../../src/ports/denylist'
+import { createClock } from '../../../src/adapters/clock'
+import { Denylist } from '../../../src/adapters/denylist'
 import { createDeployedEntitiesBloomFilter } from '../../../src/ports/deployedEntitiesBloomFilter'
 import { createNoOpDeployRateLimiter } from '../../mocks/deploy-rate-limiter-mock'
 import { createFailedDeployments } from '../../../src/ports/failedDeployments'
