+++ content/test/unit/ports/deployer.spec.ts
@@ -24,13 +24,13 @@ import * as deploymentLogic from '../../../src/logic/deployments'
 import * as deployments from '../../../src/logic/deployments'
 import { metricsDeclaration } from '../../../src/metrics'
 import { createActiveEntitiesComponent } from '../../../src/ports/activeEntities'
-import { Denylist } from '../../../src/ports/denylist'
+import { Denylist } from '../../../src/adapters/denylist'
 import { createNoOpDeployRateLimiter } from '../../mocks/deploy-rate-limiter-mock'
 import { createDeployedEntitiesBloomFilter } from '../../../src/ports/deployedEntitiesBloomFilter'
 import { createDeployer } from '../../../src/ports/deployer'
 import { createFailedDeployments } from '../../../src/ports/failedDeployments'
 import { createTestDatabaseComponent } from '../../../src/ports/postgres'
-import { createSequentialTaskExecutor } from '../../../src/ports/sequecuentialTaskExecutor'
+import { createSequentialTaskExecutor } from '../../../src/adapters/sequential-task-executor'
 import { ContentAuthenticator } from '../../../src/service/auth/Authenticator'
 import { DELTA_POINTER_RESULT } from '../../../src/service/pointers/PointerManager'
 import { EntityVersion } from '../../../src/types'
